requirements.txt
flask==2.0.1
python-dotenv==0.19.0
line-bot-sdk==2.4.1
openai==0.27.0
langchain==0.0.184
pymongo==4.1.1
python-dateutil==2.8.2
requests==2.28.1
pydantic==1.9.0
pytest==7.3.1
pytest-cov==4.0.0
black==22.3.0
flake8==4.0.1
mypy==0.991
isort==5.10.1
uvicorn==0.24.0
gunicorn==21.2.0
python-jose==3.3.0
python-multipart==0.0.6
